refactored extracting repository layer
