+++ GuildedRoseLLC/ViewController.swift
@@ -7,6 +7,7 @@ public class GreetingViewController: UIViewController, UICollectionViewDelegate 
     @IBOutlet public var noItemsLabel: UILabel!
     var data: [Item] = []
     var datasource: ItemCollectionViewDataSource?
+    
     func toggleListDisplay(data: [Item]) {
         if data.isEmpty {
             itemCollectionView.isHidden = true
@@ -16,8 +17,8 @@ public class GreetingViewController: UIViewController, UICollectionViewDelegate 
             noItemsLabel.isHidden = true
         }
     }
+    
     func setData() {
-        print("setData")
         self.datasource = ItemCollectionViewDataSource(items: self.data)
         self.toggleListDisplay(data: self.data)
         self.itemCollectionView.dataSource = self.datasource
@@ -25,19 +26,14 @@ public class GreetingViewController: UIViewController, UICollectionViewDelegate 
     
     override public func viewDidLoad() {
         super.viewDidLoad()
-
         ItemRepository().getItems() { (items: [Item]) in
             self.datasource = ItemCollectionViewDataSource(items: items)
             self.data = items
-            print("fetchData")
+            DispatchQueue.main.async {
+                self.setData()
+            }
         }
-    
-        let seconds = 5.0
-        DispatchQueue.main.asyncAfter(deadline: .now() + seconds) {
-            self.setData()
-        }
-        print("viewDidLoad")
-
+        
         self.itemCollectionView.delegate = self
         
         greeting.accessibilityIdentifier = "Greeting"
